Back-End/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
mysql-connector-python>=8.0.0
numpy>=1.24.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
python-dotenv>=1.0.0
